LP-2523 move prerequisite course helper to student app fixed tests
