+++ openedx/adg/lms/applications/constants.py
@@ -70,4 +70,3 @@ APPLICATION_REVIEW_ERROR_MSG = _('Please make a decision before submitting.')
 HTML_FOR_EMBEDDED_FILE_VIEW = '<iframe src="{path_to_file}" style="width:889px; height:393px;"></iframe>'
 
 ADG_ADMIN_GROUP_NAME = 'ADG Admins'
-DEFAULT_LANG_FOR_COURSES = 'en'
